Prevent `source ./.<shell>rc` failure
Hosts with multiple shell environments may have source files that fail when sourced under a different shell.
Currently, to enable a successful `./check-configuration.sh` on installation, `.zshrc` and `.bashrc` are sourced if they exist. So if running in Zsh, `source .bashrc` is still ran which may fail with shell-specific commands (and vice versa).

The issue is fixed by just removing it. Line 222 in this PR makes it redundant - the environment variables are exported with the `eval` statement.