Increase default `MAX_SIMULTANEOUS_VMS` config var (#72)
Done some testing and there's no disadvantage to increasing the maximum number of machines on parallel startup. It seems to use 100% of your CPU regardless of the number of VMs starting up, and memory usage only increases by the number of machines present. Setting it to `0` may be up for debate though?